+++ pennylane/_version.py
@@ -16,4 +16,4 @@
 Version number (major.minor.patch[-label])
 """
 
-__version__ = "0.28.0-dev"
+__version__ = "0.27.0"
