+++ pennylane/circuit_graph.py
@@ -18,11 +18,11 @@ representation of a quantum circuit from an Operator queue.
 # pylint: disable=too-many-branches,too-many-arguments,too-many-instance-attributes
 from collections import namedtuple
 
-import numpy as np
 import retworkx as rx
+import numpy as np
 
 import pennylane as qml
-from pennylane.measurements import MeasurementProcess, Sample, State
+
 from pennylane.wires import Wires
 
 
@@ -117,15 +117,15 @@ class CircuitGraph:
             op.queue_idx = k  # store the queue index in the Operator
 
             if hasattr(op, "return_type"):
-                if op.return_type is State:
+                if op.return_type is qml.measurements.State:
                     # State measurements contain no wires by default, but wires are
                     # required for the circuit drawer, so we recreate the state
                     # measurement with all wires
-                    op = MeasurementProcess(State, wires=wires)
+                    op = qml.measurements.MeasurementProcess(qml.measurements.State, wires=wires)
 
-                elif op.return_type is Sample and op.wires == Wires([]):
+                elif op.return_type is qml.measurements.Sample and op.wires == Wires([]):
                     # Sampling without specifying wires is treated as sampling all wires
-                    op = qml.sample(wires=wires)
+                    op = qml.measurements.MeasurementProcess(qml.measurements.Sample, wires=wires)
 
                 op.queue_idx = k
