+++ pennylane/measurements/measurements.py
@@ -690,6 +690,4 @@ class MeasurementProcess:
         new_measurement = copy.copy(self)
         if self.obs is not None:
             new_measurement.obs = self.obs.map_wires(wire_map=wire_map)
-        else:
-            new_measurement._wires = Wires([wire_map.get(wire, wire) for wire in self.wires])
         return new_measurement
