+++ pennylane/ops/functions/equal.py
@@ -16,7 +16,6 @@ This module contains the qml.equal function.
 """
 # pylint: disable=too-many-arguments,too-many-return-statements
 from typing import Union
-
 import pennylane as qml
 from pennylane.measurements import MeasurementProcess, ShadowMeasurementProcess
 from pennylane.operation import Operator
@@ -86,11 +85,11 @@ def equal(
     if op1.__class__ is not op2.__class__:
         return False
 
-    if isinstance(op1, ShadowMeasurementProcess):
-        return equal_shadow_measurements(op1, op2)
-
-    if isinstance(op1, MeasurementProcess):
+    if op1.__class__ is MeasurementProcess:
         return equal_measurements(op1, op2)
+
+    if op1.__class__ is ShadowMeasurementProcess:
+        return equal_shadow_measurements(op1, op2)
 
     return equal_operator(op1, op2, check_interface, check_trainability, rtol, atol)
