+++ pennylane/pauli/pauli_arithmetic.py
@@ -143,7 +143,7 @@ class PauliWord(dict):
     def __str__(self):
         """String representation of a PauliWord."""
         if len(self) == 0:
-            return "I"
+            return "()"
         return " @ ".join(f"{op}({w})" for w, op in self.items())
 
     def __repr__(self):
@@ -224,8 +224,6 @@ class PauliSentence(dict):
 
     def __str__(self):
         """String representation of the PauliSentence."""
-        if len(self) == 0:
-            return "I"
         return "\n+ ".join(f"{coeff} * {str(pw)}" for pw, coeff in self.items())
 
     def __repr__(self):
